chore(remix-examples): update remix examples
