+++ examples/with-remix-headless/app/routes/_auth.forgot-password.tsx
