+++ examples/with-remix-headless/app/routes/_auth.login.tsx
@@ -1,5 +1,5 @@
 import { AuthPage } from "@refinedev/core";
 
 export default function Login() {
-    return <AuthPage />;
+    return <AuthPage type="login" />;
 }
