+++ examples/with-remix-headless/app/routes/_auth.register.tsx
