--- examples/with-remix-headless/app/routes/index.tsx
@@ -1,5 +1,0 @@
-import { NavigateToResource } from "@refinedev/remix-router";
-
-export default function Index() {
-    return <NavigateToResource />;
-}
